onu_reboot_zte/src/requirements.txt
hurry.filesize
mysql-connector
requests